moved the solution to the General Reminder bug into a function so it can be reused by the Edit and the Reminders components.
